Let's avoid a big fat if
